Refactoring/renaming
