+++ src/react/components/StyleEditor/Preview.jsx
