+++ src/react/components/StyleEditor/SizeControl.jsx
