+++ src/react/components/StyleEditor/StyleFormHandler.js
