+++ src/react/components/StyleEditor/StyleSelector.jsx
