+++ src/react/components/StyleEditor/SvgRadioButton.jsx
