+++ src/react/components/StyleEditor/constants.js
@@ -1,0 +1,1 @@
+export const LOCALIZATION_BUNDLE = 'oskariui';
