+++ src/react/components/StyleEditor/index.js
@@ -1,0 +1,10 @@
+
+import * as constants from './constants';
+export { constants };
+export { AreaTab } from './AreaTab';
+export { LineTab } from './LineTab';
+export { PointTab } from './PointTab';
+export { Preview } from './Preview';
+export { SizeControl } from './SizeControl';
+export { StyleSelector } from './StyleSelector';
+export { SvgRadioButton } from './SvgRadioButton';
