+++ src/react/components/StyleEditor/lineIcons.json
