+++ src/react/index.js
@@ -28,8 +28,4 @@ export { LocalizationComponent } from './components/LocalizationComponent';
 export { Message } from './components/Message';
 export { GenericForm } from './components/GenericForm';
 export { ColorPicker } from './components/ColorPicker';
-export { SvgRadioButton } from './components/styleform/SvgRadioButton';
-export { Preview } from './components/styleform/Preview';
-export { StyleSelector } from './components/styleform/StyleSelector'
-export { SizeControl } from './components/styleform/SizeControl'
 export { StyleEditor } from './components/StyleEditor'
